App Turismo GO with IONIC
Develop of an application with the tools of IONIC to build in Android and IOS.

Description: TurismoGO is an App to enjoy traveling and discover new places and new persons.
When you are traveling, the app show you nearby places of interest around 5 kilometers. 
This is perfect if you don't know well the zone where you are.